+++ src/common/execution/request-helper.js
@@ -1,4 +1,7 @@
-const { checkWeb2SecretExists } = require('../sms/check');
+const {
+  checkWeb2SecretExists,
+  checkWeb3SecretExists,
+} = require('../sms/check');
 const { checkActiveBitInTag } = require('../utils/utils');
 const { NULL_ADDRESS, NULL_BYTES32 } = require('../utils/constant');
 const {
@@ -48,14 +51,17 @@ const checkRequestRequirements = async (
   requestorder = throwIfMissing(),
 ) => {
   await requestorderSchema().validate(requestorder);
-  const params = await createObjParams({
-    params: requestorder.params,
-    tag: requestorder.tag,
-    callback: requestorder.callback,
+  const { tag, dataset, callback, params, requester } = requestorder;
+  const paramsObj = await createObjParams({
+    params,
+    tag,
+    callback,
     noCast: true,
   });
+  const isTee = checkActiveBitInTag(tag, 1);
+
   // check encryption key
-  if (params[paramsKeyName.IEXEC_RESULT_ENCRYPTION] === true) {
+  if (paramsObj[paramsKeyName.IEXEC_RESULT_ENCRYPTION] === true) {
     const isEncryptionKeySet = await checkWeb2SecretExists(
       contracts,
       smsURL,
@@ -70,22 +76,35 @@ const checkRequestRequirements = async (
   }
   // check storage token
   if (
-    params[paramsKeyName.IEXEC_RESULT_STORAGE_PROVIDER] === 'ipfs' ||
-    params[paramsKeyName.IEXEC_RESULT_STORAGE_PROVIDER] === 'dropbox'
+    paramsObj[paramsKeyName.IEXEC_RESULT_STORAGE_PROVIDER] === 'ipfs' ||
+    paramsObj[paramsKeyName.IEXEC_RESULT_STORAGE_PROVIDER] === 'dropbox'
   ) {
     const isStorageTokenSet = await checkWeb2SecretExists(
       contracts,
       smsURL,
-      requestorder.requester,
+      requester,
       getStorageTokenKeyName(
-        params[paramsKeyName.IEXEC_RESULT_STORAGE_PROVIDER],
+        paramsObj[paramsKeyName.IEXEC_RESULT_STORAGE_PROVIDER],
       ),
     );
     if (!isStorageTokenSet) {
       throw Error(
         `Requester storage token is not set for selected provider "${
-          params[paramsKeyName.IEXEC_RESULT_STORAGE_PROVIDER]
+          paramsObj[paramsKeyName.IEXEC_RESULT_STORAGE_PROVIDER]
         }". Result archive upload will fail.`,
+      );
+    }
+  }
+  // check tee dataset encryption key
+  if (dataset && dataset !== NULL_ADDRESS && isTee) {
+    const isDatasetSecretSet = await checkWeb3SecretExists(
+      contracts,
+      smsURL,
+      dataset,
+    );
+    if (!isDatasetSecretSet) {
+      throw Error(
+        `Dataset encryption key not set for ${dataset}. Dataset decryption will fail.`,
       );
     }
   }