+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - workerpool API url configuration
 - `iexec task debug <taskid> [--logs]` to show offchain information
 - `ens.getDefaultDomain(address)` to get the default free to use ENS domain given an address
+- check dataset secret exists on requestorder check
 
 ### Changed